Support bundling with Luau types (#249)
Correctly preserve types when bundling Luau

Note that because of how Luau currently works, types that uses `typeof`
(like `type Example = typeof(variable)`) can't be hoisted correctly if
they include a local variable.

+++ src/rules/append_text_comment.rs
@@ -97,7 +97,7 @@ impl Rule for AppendTextComment {
         }
 
         let shift_lines = text.lines().count();
-        ShiftTokenLine::new(shift_lines).flawless_process(block, context);
+        ShiftTokenLine::new(shift_lines as isize).flawless_process(block, context);
 
         match self.location {
             AppendLocation::Start => {
